Bank_projekt_TUC_restructured/requirements.txt
# Database & ORM
SQLAlchemy>=2.0,<3.0
psycopg2-binary==2.9.6
alembic==1.11.1

# Data validation
pydantic>=2.10.1,<3.0.0
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1

# Workflow / Orchestration
prefect==3.4.17

# Testing
pytest==7.4.0

# Data quality
great_expectations==0.18.22

# Environment variables
python-dotenv==1.0.0

# Data analysis
pandas==2.2.3
numpy>=1.26

# Optional / useful
matplotlib==3.9.2
seaborn==0.13.2

streamlit==1.37.1

